플레이어 로직 fsm 으로 수정중, 플레이어 오른쪽 애니메이션 전부 로딩 완료

+++ Client/CAnimation.cpp
@@ -235,7 +235,7 @@ void CAnimation::PrevFrame()
 {
 	--m_iCurFrm;
 	if (0 > m_iCurFrm)
-		m_iCurFrm = m_vecFrm.size() - 1;
+		m_iCurFrm = (int)m_vecFrm.size() - 1;
 }
 
 void CAnimation::NextFrame()
